Move `unescape` to the common app
This code doesn't really have anything to do with statistics in its
design, so let's put it in a more central location.
