+++ common/utils/__init__.py
@@ -1,2 +1,3 @@
 from common.utils.paginate import *  # noqa: F403, F401
 from common.utils.echo import *  # noqa: F403, F401
+from common.utils.unescape import *  # noqa: F403, F401
