+++ statistics/utils.py
@@ -1,4 +1,3 @@
-import html
 import re
 
 from django.template import Variable, VariableDoesNotExist
@@ -33,23 +32,3 @@ def parse_kwargs(bits):
             ))
 
     return kwargs
-
-
-def unescape(content):
-    """Unescape HTML entities within template block tag delimiters
-
-    Example: the string
-    `&quot;Hello {% num_incidents categories=&#x27;1&#x27; %} worlds&quot;`
-    becomes
-    `&quot;Hello {% num_incidents categories='1'; %} worlds!&quot;`
-    """
-    tag_re = re.compile('({%.*?%})')
-    body = ''
-    position = 0
-    for match in tag_re.finditer(content):
-        start, end = match.span()
-        body += content[position:start]
-        body += html.unescape(content[start:end])
-        position = end
-    body += content[position:len(content)]
-    return body
